+++ src/components/Tree/composable/tree.js
@@ -1,4 +1,9 @@
-const supportedMode = ['single', 'multiple', 'checkbox'];
+import apiService from '../../../services/api.service';
+
+/**
+ *
+ * Convert all key value to string.
+ */
 const useStringKey = (nodes) => {
   return nodes.map((node) => {
     node.key = node.key.toString();
@@ -14,8 +19,8 @@ const useStringKey = (nodes) => {
  *
  * When extendedNode is use, it adds some options to the node properties.
  * Supported option are:
- *  - selectedColors => Tailwind utilities to applied to the selected node.
- *  - hoverColors => Tailwind utilities to be applied to a selectable node when hovered.
+ *  - collapseIcons,
+ *  - expandedIcons,
  */
 const useExtendedNode = (nodes, options) => {
   return nodes.map((node) => {
@@ -28,43 +33,29 @@ const useExtendedNode = (nodes, options) => {
   });
 };
 
-const useSelectMode = (mode, node, selectKey) => {
-  let newKey = {};
-  if (!supportedMode.includes(mode)) {
-    return newKey;
-  }
+/**
+ *
+ * Post Data request.
+ * Request will send:
+ *  - the current node key select by user,
+ *  - the mode: either select or unselect,
+ *  - an array of all key nodes selected,
+ *  - the raw value of the entire tree node,
+ *
+ */
+const usePostData = (url, action, key, selectedKeys, treeValue) => {
+  const options = {
+    method: 'POST',
+    body: JSON.stringify({ __nodeAction: action, __nodeKey: key, __nodeKeys: selectedKeys, __treeValue: treeValue }),
+  };
 
-  if (mode === 'single') {
-    newKey[node.key] = true;
-  }
-  else if (mode === 'multiple') {
-    newKey = { ...{ [node.key]: true }, ...selectKey };
-  }
-  else if (mode === 'checkbox') {
-    newKey = { ...{ [node.key]: { partialChecked: true } }, ...selectKey };
-  }
-
-  return newKey;
-};
-
-const useUnSelectMode = (mode, node, selectKey) => {
-  let newKey = {};
-  if (!supportedMode.includes(mode)) {
-    return newKey;
-  }
-
-  if (mode === 'single') {
-    newKey = {};
-  }
-  else if (mode === 'multiple') {
-    for (const [key, value] of Object.entries(selectKey)) {
-      if (key !== node.key) {
-        newKey[key] = value;
-      }
+  const { data, onFetchFinally } = apiService.fetchAsResponse(url, options);
+  onFetchFinally(() => {
+    const js = data.value?.jsRendered;
+    if (js) {
+      apiService.evalResponse(js);
     }
-  }
-
-  return newKey;
+  });
 };
 
-export { useStringKey, useExtendedNode, useSelectMode, useUnSelectMode };
+export { useStringKey, useExtendedNode, usePostData };
